ADDED: maximum from 3 float

+++ src/main/java/com/max/Maximum.java
@@ -4,14 +4,15 @@ package com.max;
 public class Maximum
 {
 
-    public Integer Max(Integer firstNumber, Integer secondNumber, Integer thirdNumber) {
-        Integer max = firstNumber;
+    public Float Max(Float firstNumber, Float secondNumber, Float thirdNumber) {
+        Float max = firstNumber;
         if (secondNumber.compareTo(max) > 0) {
             max = secondNumber;
         }
         if (thirdNumber.compareTo(max) > 0) {
             max = thirdNumber;
         }
+
         return max;
     }
